fix(rebase): update import paths after datagen sub-package restructure
After rebasing onto upstream/main, several imports still referenced the
old flat datagen layout. Update to the new sub-package paths:
- inference_perf.datagen.replay_graph_{session_datagen,types} →
  inference_perf.datagen.replay.*
- inference_perf.datagen.otel_trace_to_replay_graph →
  inference_perf.datagen.replay.*
- Remove stale synthetic_agent_sessions import (replaced by
  synthetic_agentic in the rebase)
- Regenerate docs/cli_flags.md (visionarena block ordering)
