+++ inference_perf/datagen/synthetic_agentic.py
@@ -33,8 +33,8 @@ import numpy as np
 from inference_perf.config import APIConfig, DataConfig
 from inference_perf.config.datagen.replay import SyntheticAgenticConfig
 from inference_perf.config.common import Distribution
-from inference_perf.datagen.replay_graph_session_datagen import ReplayGraphSessionGeneratorBase, ReplaySession
-from inference_perf.datagen.replay_graph_types import GraphCall, GraphEvent, InputSegment, ReplayGraph
+from inference_perf.datagen.replay.replay_graph_session_datagen import ReplayGraphSessionGeneratorBase, ReplaySession
+from inference_perf.datagen.replay.replay_graph_types import GraphCall, GraphEvent, InputSegment, ReplayGraph
 from inference_perf.datagen.synthetic_themes import (
     GENERIC_THEME,
     ROOT_SYSTEM_PROMPTS,
